+++ src/components/tabs/tabs.js
@@ -153,6 +153,7 @@ class Tabs extends React.Component {
     return {
       tabs: {
         changeValidity: this.changeValidity
+        changeWarning: this.changeWarning
       }
     };
   }
@@ -166,6 +167,14 @@ class Tabs extends React.Component {
      * @type {Object}
      */
     tabValidity: Immutable.Map()
+
+    /**
+     * Tracks the warning of each tab
+     *
+     * @property tabWarning
+     * @type {Object}
+     */
+    tabWarning: Immutable.Map()
   }
 
   /**
@@ -231,6 +240,18 @@ class Tabs extends React.Component {
   }
 
   /**
+   * Sets the warning state of the given tab (id) to the
+   * given state (warning)
+   *
+   * @method changeWarning
+   * @param {Number} id tab identifier
+   * @param {Boolean} state of tab child
+   */
+  changeWarning = (id, warning) => {
+    this.setState({ tabWarning: this.state.tabWarning.set(id, warning) });
+  }
+
+  /**
    * Handles the changing of tabs
    *
    * @method handleTabClick
@@ -270,6 +291,7 @@ class Tabs extends React.Component {
       tab.props.headerClassName,
       {
         'carbon-tabs__headers__header--error': this.state.tabValidity.get(tab.props.tabId) == false,
+        'carbon-tabs__headers__header--warning': this.state.tabWarning.get(tab.props.tabId) == true,
         'carbon-tabs__headers__header--selected': tab.props.tabId === this.state.selectedTabId
       }
     );